chore: re-organize and rename files for better readability
