+++ packages/twenty-twenty-theme/src/components/archive/archive.js
@@ -1,0 +1,51 @@
+import { connect } from "frontity";
+import React, { Fragment } from "react";
+import Article from "../post/post-item";
+import ArchiveHeader from "./archive-header";
+import Pagination from "./archive-pagination";
+import PostSeparator from "../post/post-separator";
+
+const Archive = ({ state, showExcerpt, showMedia }) => {
+  // Get the data of the current list.
+  const data = state.source.get(state.router.link);
+  const { primary } = state.theme.colors;
+
+  return (
+    <>
+      {/* If the list is a taxonomy, we render a title. */}
+      {data.isTaxonomy && (
+        <ArchiveHeader labelColor={primary} label={data.taxonomy}>
+          <span>{state.source[data.taxonomy][data.id].name}</span>
+        </ArchiveHeader>
+      )}
+
+      {/* If the list is for a specific author, we render a title. */}
+      {data.isAuthor && (
+        <ArchiveHeader labelColor={primary} label="Author">
+          <b>{state.source.author[data.id].name}</b>
+        </ArchiveHeader>
+      )}
+
+      {/* Iterate over the items of the list. */}
+      {data.items.map(({ type, id }, index) => {
+        const isLastArticle = index === data.items.length - 1;
+        const item = state.source[type][id];
+        // Render one Item component for each one.
+        return (
+          <Fragment key={item.id}>
+            <Article
+              key={item.id}
+              item={item}
+              showExcerpt={showExcerpt}
+              showMedia={showMedia}
+            />
+            {!isLastArticle && <PostSeparator />}
+          </Fragment>
+        );
+      })}
+      <Pagination />
+    </>
+  );
+};
+
+export default connect(Archive);
