+++ packages/twenty-twenty-theme/src/components/mobile/menu-button.js
@@ -1,7 +1,11 @@
 import { connect } from "frontity";
 import React from "react";
 import { ToggleIcon } from "../icons";
-import { LabeledIcon, NavToggle, ToggleWrapper } from "../nav-toggle";
+import {
+  LabeledIcon,
+  NavToggle,
+  ToggleWrapper
+} from "../navigation/nav-toggle";
 
 const MobileMenuButton = ({ state, actions }) => {
   // Get the menu state and action
