+++ packages/twenty-twenty-theme/src/components/post/post-separator.js
