+++ packages/twenty-twenty-theme/src/components/post/post-tags.js
