+++ packages/twenty-twenty-theme/src/components/search/search-button.js
@@ -1,7 +1,11 @@
 import { connect, styled } from "frontity";
 import React from "react";
 import { SearchIcon } from "../icons";
-import { BaseToggle, LabeledIcon, ToggleWrapper } from "../nav-toggle";
+import {
+  BaseToggle,
+  LabeledIcon,
+  ToggleWrapper
+} from "../navigation/nav-toggle";
 
 const SearchButton = ({ state, actions }) => {
   // Get the state of the search modal
